+++ package.json
@@ -5,7 +5,8 @@
   "main": "dist/index.js",
   "scripts": {
     "build": "rollup -c",
-    "test": "find ./test -name '*.spec.ts' | xargs mocha "
+    "test": "find ./test -name '*.spec.ts' | xargs mocha ",
+    "test2": "jest"
   },
   "keywords": [
     "json",
@@ -26,17 +27,20 @@
     "typescript": "^5.4.5"
   },
   "devDependencies": {
+    "@babel/core": "^7.24.4",
+    "@babel/preset-env": "^7.24.4",
+    "@babel/preset-typescript": "^7.24.1",
     "@eslint/js": "^9.1.1",
+    "@jest/globals": "^29.7.0",
     "@types/eslint__js": "^8.42.3",
     "@types/node": "^20.12.7",
     "@types/stream-array": "^1.1.4",
     "@typescript-eslint/eslint-plugin": "^7.7.1",
     "@typescript-eslint/parser": "^7.7.1",
-    "chai": "^3.5.0",
-    "chai-as-promised": "^5.3.0",
+    "babel-jest": "^29.7.0",
     "eslint": "^8.57.0",
     "globals": "^15.0.0",
-    "mocha": "^10.4.0",
+    "jest": "^29.7.0",
     "prettier": "^3.2.5",
     "typescript-eslint": "^7.7.1"
   }
